+++ node_helper.js
@@ -1,28 +1,97 @@
-/* Magic Mirror
- * Module: MMM-NHL
+/**
+ * @file node_helper.js
  *
- * By fewieden https://github.com/fewieden/MMM-NHL
- * MIT Licensed.
+ * @author fewieden
+ * @license MIT
+ *
+ * @see  https://github.com/fewieden/MMM-NHL
  */
 
 /* eslint-env node */
 
+/**
+ * @external node-fetch
+ * @see https://www.npmjs.com/package/node-fetch
+ */
 const fetch = require('node-fetch');
+
+/**
+ * @external querystring
+ * @see https://nodejs.org/api/querystring.html
+ */
 const qs = require('querystring');
+
+/**
+ * @external node_helper
+ * @see https://github.com/MichMich/MagicMirror/blob/master/js/node_helper.js
+ */
 const NodeHelper = require('node_helper');
 
 const BASE_URL = 'https://statsapi.web.nhl.com/api/v1';
 
+/**
+ * Derived team details of a game from API endpoint for easier usage.
+ *
+ * @typedef {object} Team
+ * @property {number} id - Team identifier.
+ * @property {string} name - Full team name.
+ * @property {string} short - 3 letter team name.
+ * @property {number} score - Current score of the team.
+ */
+
+/**
+ * Derived game details from API endpoint for easier usage.
+ *
+ * @typedef {object} Game
+ * @property {number} id - Game identifier.
+ * @property {string} timestamp - Start date of the game in UTC timezone.
+ * @property {object} status - Contains information about the game status.
+ * @property {string} status.abstract - Abstract game status e.g. Preview, Live or Final.
+ * @property {string} status.detailed - More detailed version of the abstract game status.
+ * @property {object} teams - Contains information about both teams.
+ * @property {Team} teams.away - Contains information about the away team.
+ * @property {Team} teams.home - Contains information about the home team.
+ * @property {object} live - Contains information about the live state of the game.
+ * @property {string} live.period - Period of the game e.g. 1st, 2nd, 3rd, OT or SO.
+ * @property {string} live.timeRemaining - Remaining time of the current period in format mm:ss.
+ */
+
+/**
+ * Derived season details from API endpoint for easier usage.
+ *
+ * @typedef {object} SeasonDetails
+ * @property {string} year - Year of the season in format yy/yy e.g. 20/21.
+ * @property {string} mode - Mode of the season e.g. PR, R and P.
+ */
+
+/**
+ * @module node_helper
+ * @description Backend for the module to query data from the API provider.
+ *
+ * @requires external:node-fetch
+ * @requires external:querystring
+ * @requires external:node_helper
+ */
 module.exports = NodeHelper.create({
-    games: [],
-    season: {},
+    /** @member {?Game} nextGame - The next upcoming game is stored in this variable. */
     nextGame: null,
+    /** @member {Game[]} liveGames - List of all ongoing games. */
     liveGames: [],
 
+    /**
+     * @function socketNotificationReceived
+     * @description Receives socket notifications from the module.
+     * @async
+     * @override
+     *
+     * @param {string} notification - Notification name
+     * @param {*} payload - Detailed payload of the notification.
+     *
+     * @returns {void}
+     */
     async socketNotificationReceived(notification, payload) {
         if (notification === 'CONFIG') {
             this.config = payload.config;
-            this.teams = payload.teams;
 
             await this.initTeams();
 
@@ -32,6 +101,13 @@ module.exports = NodeHelper.create({
         }
     },
 
+    /**
+     * @function initTeams
+     * @description Retrieves a list of all teams from the API and initializes teamMapping.
+     * @async
+     *
+     * @returns {void}
+     */
     async initTeams() {
         const response = await fetch(`${BASE_URL}/teams`);
 
@@ -48,10 +124,15 @@ module.exports = NodeHelper.create({
 
             return mapping;
         }, {});
-
-        this.sendSocketNotification('TEAM_MAPPING', this.teamMapping);
     },
 
+    /**
+     * @function fetchSchedule
+     * @description Retrieves a list of games from the API with timespan based on config options.
+     * @async
+     *
+     * @returns {object[]} Raw games from API endpoint.
+     */
     async fetchSchedule() {
         const date = new Date();
         date.setDate(date.getDate() - this.config.daysInPast);
@@ -73,6 +154,12 @@ module.exports = NodeHelper.create({
         return dates.map(date => date.games).flat();
     },
 
+    /**
+     * @function filterGameByFocus
+     * @description Helper function to filter games based on config option.
+     *
+     * @returns {boolean} Should game remain in list?
+     */
     filterGameByFocus(game) {
         const focus = this.config.focus_on;
         if (!focus) {
@@ -85,6 +172,14 @@ module.exports = NodeHelper.create({
         return focus.includes(homeTeam) || focus.includes(awayTeam);
     },
 
+    /**
+     * @function computeSeasonDetails
+     * @description Computes current season details (year and mode) from list of games.
+     *
+     * @param {object[]} schedule - List of raw games from API endpoint.
+     *
+     * @returns {SeasonDetails} Current season details.
+     */
     computeSeasonDetails(schedule) {
         const game = schedule.find(game => game.status.abstractGameState !== 'Final') || schedule[schedule.length - 1];
 
@@ -105,6 +200,15 @@ module.exports = NodeHelper.create({
         };
     },
 
+    /**
+     * @function parseTeam
+     * @description Transforms raw team information for easier usage.
+     *
+     * @param {object} teams - Both teams in raw format.
+     * @param {string} type - Type of team: away or home.
+     *
+     * @returns {Team} Parsed team information.
+     */
     parseTeam(teams = {}, type) {
         return {
             id: teams[type].team.id,
@@ -114,6 +218,14 @@ module.exports = NodeHelper.create({
         };
     },
 
+    /**
+     * @function parseGame
+     * @description Transforms raw game information for easier usage.
+     *
+     * @param {object} game - Raw game information.
+     *
+     * @returns {Game} Parsed game information.
+     */
     parseGame(game = {}) {
         return {
             id: game.gamePk,
@@ -133,31 +245,63 @@ module.exports = NodeHelper.create({
         };
     },
 
-    setNextGame(games) {
-        this.nextGame = games.find(game => game?.status?.abstract === 'Preview');
-        this.liveGames = games.filter(game => game?.status?.abstract === 'Live');
+    /**
+     * @function setNextandLiveGames
+     * @description Sets the next scheduled and live games from a list of games.
+     *
+     * @param {Game[]} games - List of games.
+     *
+     * @returns {void}
+     */
+    setNextandLiveGames(games) {
+        this.nextGame = games.find(game => game.status.abstract === 'Preview');
+        this.liveGames = games.filter(game => game.status.abstract === 'Live');
     },
 
-    sort(game1, game2) {
+    /**
+     * @function sortGamesByTimestampAndID
+     * @description Helper function to sort games by timestamp and ID.
+     *
+     * @param {object} game1 - Raw game information of first game.
+     * @param {object} game2 - Raw game information of second game.
+     *
+     * @returns {number} Should game be before or after in the list?
+     */
+    sortGamesByTimestampAndID(game1, game2) {
         if (game1.gameDate === game2.gameDate) {
             return game1.id > game2.id ? 1 : -1;
         }
+
         return game1.gameDate > game2.gameDate ? 1 : -1;
     },
 
+    /**
+     * @function updateSchedule
+     * @description Retrieves new schedule from API and sends a socket notification to the module.
+     * @async
+     *
+     * @returns {void}
+     */
     async updateSchedule() {
         const schedule = await this.fetchSchedule();
-        schedule.sort(this.sort);
+        schedule.sort(this.sortGamesByTimestampAndID);
         const season = this.computeSeasonDetails(schedule);
 
         const focusSchedule = schedule.filter(this.filterGameByFocus.bind(this));
 
         const games = focusSchedule.map(this.parseGame.bind(this));
 
-        this.setNextGame(games);
+        this.setNextandLiveGames(games);
         this.sendSocketNotification('SCHEDULE', {games, season});
     },
 
+    /**
+     * @function fetchOnLiveState
+     * @description If there is a live game trigger updateSchedule.
+     * @async
+     *
+     * @returns {void}
+     */
     fetchOnLiveState() {
         const hasLiveGames = this.liveGames.length > 0;
         const gameAboutToStart = this.nextGame && new Date() > new Date(this.nextGame.timestamp);